VisionCare/requirements.txt
fastapi
pydantic
groq
python-dotenv
PyMuPDF
python-docx
uvicorn[standard]